Collapsed `Handle` tests and fixed Scala 3 issues
